- remove casting, thanks to types now being well-defined
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@2765 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/s2map.cc
@@ -580,10 +580,10 @@ void S2_Map_Loader::load_s2mf(Editor_Game_Base *game)
 				default:   res = "";       amount = 0;
 				}
 
-            uint8_t nres=0;
+            int32_t nres=0;
             if (res!="") {
                nres=m_map->get_world()->get_resource(res.c_str());
-               if (static_cast<uint8_t>(nres)==-1)
+               if (nres==-1)
                   throw wexception("World doesn't define Resource %s\n, you can't play settler maps here!\n", res.c_str());
 				}
 				const int32_t real_amount = static_cast<int32_t>
